# src/modules/voucher/dto/create-voucher.dto.ts
import { IsDateString, IsNotEmpty, IsNumber, IsUUID } from 'class-validator';

export class CreateVoucherDto {
  @IsNotEmpty()
  @IsNumber()
  specialOfferId: number;

  @IsDateString()
  @IsNotEmpty()
  expirationDate: string;
}